+++ src/controls/actions.ts
@@ -19,7 +19,7 @@ import { skewCursorStyleHandler, skewHandlerX, skewHandlerY } from './skew';
 import { getLocalPoint, getActionFromCorner } from './util';
 import { wrapWithFireEvent } from './wrapWithFireEvent';
 import { wrapWithFixedAnchor } from './wrapWithFixedAnchor';
-import { PolyControl } from './polycontrol.class'
+import { PolyControl } from './PolyControl'
 
 export {
   scaleCursorStyleHandler,
